+++ internal_testsuites/golang/testsuite/startosis_define_fact_wait_test/startosis_define_fact_wait_test.go
@@ -67,7 +67,7 @@ func TestStartosis(t *testing.T) {
 	logrus.Infof("Executing Startosis script...")
 	logrus.Debugf("Startosis script content: \n%v", startosisScript)
 
-	outputStream, _, err := enclaveCtx.ExecuteKurtosisScript(ctx, startosisScript, defaultDryRun)
+	outputStream, _, err := enclaveCtx.RunStarlarkScript(ctx, startosisScript, defaultDryRun)
 	require.NoError(t, err, "Unexpected error executing startosis script")
 	scriptOutput, _, interpretationError, validationErrors, executionError := test_helpers.ReadStreamContentUntilClosed(outputStream)
 
@@ -81,9 +81,9 @@ Service dependency 2 deployed successfully.
 	require.Equal(t, expectedScriptOutput, scriptOutput)
 	logrus.Infof("Successfully ran Startosis script")
 
-	servicesSet, err := enclaveCtx.GetServices()
-	require.Equal(t, 3, len(servicesSet))
-	require.Contains(t, servicesSet, services.ServiceID(expectedGetOutput))
-	require.Contains(t, servicesSet, services.ServiceID(expectedPostOutput))
+	serviceInfos, err := enclaveCtx.GetServices()
+	require.Equal(t, 3, len(serviceInfos))
+	require.Contains(t, serviceInfos, services.ServiceID(expectedGetOutput))
+	require.Contains(t, serviceInfos, services.ServiceID(expectedPostOutput))
 	require.Nil(t, err)
 }