+++ internal_testsuites/golang/testsuite/startosis_module_test/shared_constants_test.go
@@ -1,7 +1,0 @@
-package startosis_module_test
-
-const (
-	isPartitioningEnabled = false
-	emptyExecuteParams    = "{}"
-	defaultDryRun         = false
-)