+++ internal_testsuites/golang/testsuite/startosis_package_test/shared_constants_test.go
@@ -1,0 +1,7 @@
+package startosis_package_test
+
+const (
+	isPartitioningEnabled = false
+	emptyRunParams        = "{}"
+	defaultDryRun         = false
+)